+++ modules/canvaskit/skottie.js
@@ -10,7 +10,11 @@
 
 // soundMap is an optional object that maps string names to AudioPlayers
 // AudioPlayers manage a single audio layer with a seek function
-CanvasKit.MakeManagedAnimation = function(json, assets, prop_filter_prefix, soundMap) {
+
+// logger is an optional logging object, expected to provide two functions:
+//   - onError(err_str, json_node_str)
+//   - onWarning(wrn_str, json_node_str)
+CanvasKit.MakeManagedAnimation = function(json, assets, prop_filter_prefix, soundMap, logger) {
   if (!CanvasKit._MakeManagedAnimation) {
     throw 'Not compiled with MakeManagedAnimation';
   }
@@ -18,7 +22,8 @@ CanvasKit.MakeManagedAnimation = function(json, assets, prop_filter_prefix, soun
     prop_filter_prefix = '';
   }
   if (!assets) {
-    return CanvasKit._MakeManagedAnimation(json, 0, nullptr, nullptr, nullptr, prop_filter_prefix, soundMap);
+    return CanvasKit._MakeManagedAnimation(json, 0, nullptr, nullptr, nullptr, prop_filter_prefix,
+                                           soundMap, logger);
   }
   var assetNamePtrs = [];
   var assetDataPtrs = [];
@@ -52,7 +57,8 @@ CanvasKit.MakeManagedAnimation = function(json, assets, prop_filter_prefix, soun
   var assetSizesPtr = copy1dArray(assetSizes,    "HEAPU32");
 
   var anim = CanvasKit._MakeManagedAnimation(json, assetKeys.length, namesPtr,
-                                             assetsPtr, assetSizesPtr, prop_filter_prefix, soundMap);
+                                             assetsPtr, assetSizesPtr, prop_filter_prefix,
+                                             soundMap, logger);
 
   // The C++ code has made copies of the asset and string data, so free our copies.
   CanvasKit._free(namesPtr);